requirements/base.txt
python-telegram-bot==20.7
emoji==2.10.1
boto3==1.34.34